feat/fe/shareadd: shareAdd shareEdit 공통 컴포넌트 구현
